+++ crates/cli/tests/reference/targets-target-web-experimental-reset-state-function-atomics.js
@@ -5,6 +5,7 @@ export function __wbg_reset_state () {
     cachedUint8ArrayMemory0 = null;
     if (typeof numBytesDecoded !== 'undefined') numBytesDecoded = 0;
 
+    __wbg_called_abort = false;
     const wasmInstance = new WebAssembly.Instance(wasmModule, __wbg_get_imports());
     wasm = wasmInstance.exports;
     wasm.__wbindgen_start();